+++ src/main/java/pyre/coloredredstone/items/IColoredItem.java
@@ -1,10 +1,21 @@
 package pyre.coloredredstone.items;
 
+import net.minecraft.advancements.CriteriaTriggers;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.item.EntityItem;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
+import net.minecraft.init.MobEffects;
+import net.minecraft.init.SoundEvents;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.potion.PotionEffect;
+import net.minecraft.stats.StatList;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.EnumActionResult;
+import net.minecraft.util.EnumHand;
+import net.minecraft.util.SoundCategory;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.ReflectionHelper;
@@ -19,6 +30,9 @@ import java.lang.reflect.Field;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+
+import static pyre.coloredredstone.blocks.IColoredFeatures.BURNABLE_COLOR;
+import static pyre.coloredredstone.blocks.IColoredFeatures.EDIBLE_COLOR;
 
 public interface IColoredItem {
 
@@ -105,6 +119,11 @@ public interface IColoredItem {
                     tooltips.add(chatColor + ColoredRedstone.proxy.localize(Reference.MOD_ID + ".tooltip.fishy"));
                 }
                 break;
+            case MAGENTA:
+                if (CurrentModConfig.edible) {
+                    tooltips.add(chatColor + ColoredRedstone.proxy.localize(Reference.MOD_ID + ".tooltip.edible"));
+                }
+                break;
             case ORANGE:
                 if (CurrentModConfig.explosionproof) {
                     tooltips.add(chatColor + ColoredRedstone.proxy.localize(Reference.MOD_ID + ".tooltip.explosionproof"));
@@ -139,9 +158,35 @@ public interface IColoredItem {
     }
 
     default int getBurnTime(ItemStack itemStack) {
-        if (CurrentModConfig.burnable && itemStack.getMetadata() == EnumColor.BROWN.getMetadata()) {
+        if (CurrentModConfig.burnable && itemStack.getMetadata() == BURNABLE_COLOR.getMetadata()) {
             return CurrentModConfig.burnableBurningTime;
         }
         return 0;
+    }
+
+    default ActionResult<ItemStack> tryEatItem(World worldIn, EntityPlayer playerIn, EnumHand handIn) {
+        if (CurrentModConfig.edible && playerIn.isSneaking() && playerIn.canEat(false)) {
+            ItemStack heldItem = playerIn.getHeldItem(handIn);
+            EnumColor color = EnumColor.byMetadata(heldItem.getMetadata());
+            if (color == EDIBLE_COLOR) {
+                playerIn.getFoodStats().addStats(CurrentModConfig.edibleHunger, CurrentModConfig.edibleSaturation);
+                worldIn.playSound(null, playerIn.posX, playerIn.posY, playerIn.posZ, SoundEvents.ENTITY_PLAYER_BURP,
+                        SoundCategory.PLAYERS, 0.5F, worldIn.rand.nextFloat() * 0.1F + 0.9F);
+                if (!worldIn.isRemote && worldIn.rand.nextInt(100) < CurrentModConfig.edibleSaturationChance) {
+                    playerIn.addPotionEffect(new PotionEffect(MobEffects.SATURATION, CurrentModConfig.edibleSaturationDuration));
+                }
+                if (!worldIn.isRemote && worldIn.rand.nextInt(100) + 1 < CurrentModConfig.edibleNauseaChance) {
+                    playerIn.addPotionEffect(new PotionEffect(MobEffects.NAUSEA, CurrentModConfig.edibleNauseaDuration));
+                }
+                playerIn.addStat(StatList.getObjectUseStats(heldItem.getItem()));
+
+                if (playerIn instanceof EntityPlayerMP) {
+                    CriteriaTriggers.CONSUME_ITEM.trigger((EntityPlayerMP) playerIn, heldItem);
+                }
+                heldItem.shrink(1);
+                return new ActionResult<>(EnumActionResult.SUCCESS, playerIn.getHeldItem(handIn));
+            }
+        }
+        return new ActionResult<>(EnumActionResult.FAIL, playerIn.getHeldItem(handIn));
     }
 }