+++ src/main/java/pyre/coloredredstone/items/ItemColoredRedstoneDust.java
@@ -7,16 +7,18 @@ import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
+import net.minecraft.init.MobEffects;
+import net.minecraft.init.SoundEvents;
 import net.minecraft.item.ItemStack;
-import net.minecraft.util.EnumActionResult;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.EnumHand;
-import net.minecraft.util.NonNullList;
+import net.minecraft.potion.PotionEffect;
+import net.minecraft.stats.StatList;
+import net.minecraft.util.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import pyre.coloredredstone.blocks.BlockColoredRedstoneWire;
+import pyre.coloredredstone.config.CurrentModConfig;
 import pyre.coloredredstone.init.ModBlocks;
 import pyre.coloredredstone.util.EnumColor;
 
@@ -89,5 +91,10 @@ public class ItemColoredRedstoneDust extends ItemBase implements IColoredItem {
     @Override
     public int getItemBurnTime(ItemStack itemStack) {
         return getBurnTime(itemStack);
+    }
+
+    @Override
+    public ActionResult<ItemStack> onItemRightClick(World worldIn, EntityPlayer playerIn, EnumHand handIn) {
+        return tryEatItem(worldIn, playerIn, handIn);
     }
 }