+++ src/main/java/pyre/coloredredstone/items/ItemColoredRedstoneLamp.java
@@ -4,8 +4,11 @@ import net.minecraft.block.Block;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.Entity;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemColored;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.EnumHand;
 import net.minecraft.util.NonNullList;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
@@ -60,6 +63,11 @@ public class ItemColoredRedstoneLamp extends ItemColored implements IColoredItem
     @Override
     public int getItemBurnTime(ItemStack itemStack) {
         return getBurnTime(itemStack);
+    }
+
+    @Override
+    public ActionResult<ItemStack> onItemRightClick(World worldIn, EntityPlayer playerIn, EnumHand handIn) {
+        return tryEatItem(worldIn, playerIn, handIn);
     }
 
     @SideOnly(Side.CLIENT)