+++ src/main/java/pyre/coloredredstone/items/ItemColoredRedstoneTorch.java
@@ -88,4 +88,9 @@ public class ItemColoredRedstoneTorch extends ItemColored implements IColoredIte
     public int getItemBurnTime(ItemStack itemStack) {
         return getBurnTime(itemStack);
     }
+
+    @Override
+    public ActionResult<ItemStack> onItemRightClick(World worldIn, EntityPlayer playerIn, EnumHand handIn) {
+        return tryEatItem(worldIn, playerIn, handIn);
+    }
 }