+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 - New colored property - Alienated - cyan blocks have a chance to drop Enderman loot instead of actual block item. Default: 5% chance to drop.
 - New colored property - Fishy - light blue blocks have a chance to drop fishing loot instead of actual block item. Default: 5% chance to drop.
 - New colored property - Soft - purple blocks reduce fall damage. Default: 50% fall damage reduction.
+- New colored property - Edible - magenta items can be eaten. To eat an item press Shift + Right Click (**Placing item/block has priority over eating!**). By default eating magenta item restores 2 points of Hunger and 0.3 points of Saturation. There is also 5% chance to get Saturation effect for 1 second and/or Nausea effect for 15 seconds.
 - All of new colored properties can be enabled/disabled in mod config. Chance and specific parameters of properties can also be customized in config.
+### Changed
+- Small changes to English config translation.
 ## 1.2.1
 ### Fixed
 - Added missing entry in polish translation.